refactor(theme): share hex color normalization helper
- add logic/hexColor normalizeHexColor utility

- remove duplicated HEX_COLOR_REGEX and normalizeHexColor in theme/index and useUserPreferences

- keep accent validation behavior identical across preferences and theme resolution
